Check latexmk logs for common errors

Check latexmk logs for common errors

Usage

check_log(slide_file, before = 0, after = 1)

Arguments

slide_file

Name of a slide file, e.g. "slides-gp-basic-3", with optional file extension.

before, after

[integer(1)] Number of log lines to display before and after the line found via regex. Defaults to 0 lines before, 1 line after.

Value

A character with one element per match, with individual lines separated by \\n within each element.

Examples

if (FALSE) { # \dontrun{
check_log("slides-gp-basic-3")
# "976: ! Undefined control sequence.\n976: l.9 \newcommandiscrete\n"
} # }