Run make in a slide directory

Run make in a slide directory

Usage

make_slides(
  topic,
  lectures_tbl = collect_lectures(),
  make_arg = "most",
  pre_clean = TRUE,
  check_status = TRUE,
  verbose = TRUE,
  log = FALSE
)

Arguments

topic

character(1)> Topic sub-directory to run make in, e.g. slides-cart-predictions

lectures_tbl

Must contain tex column. Defaults to collect_lectures().

make_arg

["most"]: Likely not wise to change. make all also copies slides to slides-pdf, which may not be intended.

pre_clean

[TRUE]: Run make clean beforehand, ensuring a clean slate.

check_status

[TRUE]: Wait for make to finish and return the exit status.

verbose

[TRUE]: Print additional output to the console.

log

[FALSE]: Write stdout and stderr logs to ./logs/.

Value

Invisibly: A list with entries

passed: TRUE indicates a successful comparison, FALSE a failure.

log: Absolute path to the log file in case of a non-zero exit status.

Examples

if (FALSE) { # \dontrun{
# Default: Run make and output status
make_slides("cart")
make_slides("regularization", check_status = TRUE)

# Runs in background, but doesn't capture exit code
make_slides("cart", check_status = FALSE)
} # }